+++ doc/changelog.rst
@@ -12,6 +12,8 @@ Version 0.9.7
 * Rename ``--name=NAME`` option to ``--benchmark=NAME``
 * Remove ``perf.monotonic_clock()`` since it wasn't monotonic on Python 2.7.
 * Remove ``is_significant()`` from the public API
+* Run warmup values must now be normalized per loop iteration.
+* dump command now displays raw value of calibration runs.
 
 Version 0.9.6 (2017-03-15)
 --------------------------